google-sheets-mcp-server/pyproject.toml
[project]
name = "google-sheets-mcp-server"
version = "0.1.0"
description = "MCP server for Google Sheets"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.47.0",
    "google-auth-oauthlib>=1.2.3",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
